クリックが正常に反応しなかった問題を解決した｡ タイムオーバーの時に延々にデッドカウントが加算されていた問題を解決した｡ 操作をレコーディングするクラスを作成した｡

+++ BomberProject/Factory_CheckPoint.cpp
@@ -262,10 +262,10 @@ void	CheckPoint::PlayerPass(UpdatePacket &i_UpdatePacket){
 ////            F
 ////
 CheckPoint::WORKSTATE CheckPoint::Reduction(UpdatePacket &i_UpdatePacket){
-	m_vLineScale.x		-= m_fBassLength / m_fReductionTime * i_UpdatePacket.GetTime()->getElapsedTime() ;
+	m_vLineScale.x		-= m_fBassLength / m_fReductionTime * (float)i_UpdatePacket.GetTime()->getElapsedTime() ;
 	if(!ActiveIsLast()){
-		m_vStringScale.x	-= m_fBassLength / 4 / m_fReductionTime * i_UpdatePacket.GetTime()->getElapsedTime() ;
-		m_vStringScale.y	-= 2 / m_fReductionTime * i_UpdatePacket.GetTime()->getElapsedTime() ;
+		m_vStringScale.x	-= m_fBassLength / 4 / m_fReductionTime * (float)i_UpdatePacket.GetTime()->getElapsedTime() ;
+		m_vStringScale.y	-= 2 / m_fReductionTime * (float)i_UpdatePacket.GetTime()->getElapsedTime() ;
 	}
 	if( m_vLineScale.x >= 0.0f ){
 		return WORKSTATE_UNFINSHED ;
